fixed: loading the first timesteps would fail due to the use of the unsigned counter
git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@2757 e10b68d5-8a6e-419e-a041-bce267b0401d

+++ Apps/Common/AppCommon.h
@@ -39,7 +39,7 @@ namespace SIM
     reader.registerWriter(hdf);
     int max = reader.getTimeLevel();
     solver.fastForward(reader.realTimeLevel(max-steps));
-    for (size_t i=steps;i>=0;--i) {
+    for (int i=steps;i>=0;--i) {
       reader.loadTimeLevel(max-i,xml,hdf);
       solver.postSolve(solver.getTimePrm());
       solver.advanceStep();
